move github summary feature into integration app
