+++ tests/apps/github_summary/task_test.py
@@ -4,14 +4,14 @@ import pytest
 
 from busy_beaver.adapters.slack import Channel
 from busy_beaver.models import ApiUser
-from busy_beaver.apps.github_summary.task import (
+from busy_beaver.apps.github_integration.summary.task import (
     start_post_github_summary_task,
     fetch_github_summary_post_to_slack,
 )
 from busy_beaver.toolbox import utc_now_minus
 from tests._utilities import FakeSlackClient
 
-MODULE_TO_TEST = "busy_beaver.apps.github_summary.task"
+MODULE_TO_TEST = "busy_beaver.apps.github_integration.summary.task"
 
 
 @pytest.fixture
